backend/requirements.txt
# =====================================================
# Install PyTorch separately for your CUDA version.
# Example (CUDA 12.1+):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# =====================================================

fastapi>=0.110.0
uvicorn>=0.23.0
whisperlivekit>=0.2.12
transformers>=4.36.0
sentencepiece>=0.1.99
ollama>=0.1.0
openai>=1.0.0
pyyaml>=6.0
